Adding pki.ldapdb module
