+++ base/server/python/pki/server/__init__.py
@@ -39,6 +39,8 @@ import pki
 import pki.nssdb
 import pki.util
 import six
+from pki.ldapdb import PKILDAPDatabase
+
 from lxml import etree
 
 INSTANCE_BASE_DIR = '/var/lib/pki'
@@ -387,7 +389,7 @@ class PKISubsystem(object):
                 'Invalid parameter value in %s.ldapconn.secureConn: %s' %
                 (name, secure))
 
-        connection = PKIDatabaseConnection(url)
+        connection = PKILDAPDatabase(url)
 
         connection.set_security_database(self.instance.nssdb_dir)
 
@@ -990,64 +992,6 @@ class PKIInstance(object):
         if self.type == 9:
             return "Dogtag 9 " + self.name
         return self.name
-
-
-class PKIDatabaseConnection(object):
-
-    def __init__(self, url='ldap://localhost:389'):
-
-        self.url = url
-
-        self.nssdb_dir = None
-
-        self.bind_dn = None
-        self.bind_password = None
-
-        self.client_cert_nickname = None
-        self.nssdb_password = None
-
-        self.temp_dir = None
-        self.ldap = None
-
-    def set_security_database(self, nssdb_dir=None):
-        self.nssdb_dir = nssdb_dir
-
-    def set_credentials(self, bind_dn=None, bind_password=None,
-                        client_cert_nickname=None, nssdb_password=None):
-        self.bind_dn = bind_dn
-        self.bind_password = bind_password
-        self.client_cert_nickname = client_cert_nickname
-        self.nssdb_password = nssdb_password
-
-    def open(self):
-
-        self.temp_dir = tempfile.mkdtemp()
-
-        if self.nssdb_dir:
-
-            ldap.set_option(ldap.OPT_X_TLS_CACERTDIR, self.nssdb_dir)
-
-        if self.client_cert_nickname:
-
-            password_file = os.path.join(self.temp_dir, 'password.txt')
-            with open(password_file, 'w') as f:
-                f.write(self.nssdb_password)
-
-            ldap.set_option(ldap.OPT_X_TLS_CERTFILE, self.client_cert_nickname)
-            ldap.set_option(ldap.OPT_X_TLS_KEYFILE, password_file)
-
-        self.ldap = ldap.initialize(self.url)
-
-        if self.bind_dn and self.bind_password:
-            self.ldap.simple_bind_s(self.bind_dn, self.bind_password)
-
-    def close(self):
-
-        if self.ldap:
-            self.ldap.unbind_s()
-
-        if self.temp_dir:
-            shutil.rmtree(self.temp_dir)
 
 
 class PKIServerException(pki.PKIException):
